+++ mypy_pure/tests/resources/pure_calls_external_impure.py
@@ -1,4 +1,4 @@
-import external_module
+import external_module  # type: ignore[import-not-found]
 
 from mypy_pure.decorators import pure